Color blindness simulation for kmag... not so useful for people with color blindness, but good for designing accessible interfaces (also makes a fun toy in its own right).
Several algorithms are implemented (the preprocessor selects one at compile time), so some of the evolution is visible. I'd like to get the "non-simple" algorithm working, but the paper is missing most of the crucial variables. For now the "simple" algorithm with coefficients derived from Visolve is used.

TODO: make gamma configurable, re-add 'invert colors', add modes to actually help people with color blindness in addition to the simulation modes... UI probably needs more work to make this practical, the simple menu-with-list-of-exclusive-options doesn't really cut it.

svn path=/trunk/KDE/kdeaccessibility/kmag/; revision=774789