Unbreak screen grabbing
After 39f712a2, KMag would only show small parts of the screen and its
view was overlayed with black and gray areas and even image recursion.
The cursor position would also not correspond to its actual position.

This is due to changing `QApplication::desktop()->winId()` to the window
id of KMag, which is wrong when the whole screen should be grabbed
instead.

Using `0` for the window id is interpreted as the root window by Qt and
fixes the issue.

Test Plan:

Move cursor to the top left and top right corner of the screen. KMag
should always magnify the correct area without any black bars, and the
cursor position should be correctly indicated too.